requirements.txt
tensorflow
Pillow
matplotlib
scikit-learn
ipywidgets
tqdm
numpy
